requirements/base.txt
Django==1.6
Paste==1.7.5.1
Pygments==1.6
South==0.8.3
Werkzeug==0.9.4
boto==2.17.0
dj-database-url==0.2.2
django-braces==1.2.2
django-celery==3.1.1
django-model-utils==1.5.0
django-bootstrap-form==3.0
gevent==0.13.8
logutils==0.3.3
psycopg2==2.5.1
requests==2.0.1
static==0.4
gunicorn==18.0
stripe==1.9.8
